script: add if statement to prevent garbage output
The command `playerctl metadata` will print `No player data found` when
I am not playing songs. And this will cause a list of error message in
the log. So here I add a if statement to see if the playerctl return 0
exit code. If true then get the metadata. Else it will print `No songs
playing` and exit.

Also I've moved foreground color at the front of the file that user can
easily modify the value.

Signed-off-by: Avimitin <avimitin@gmail.com>